feat: rearrange the folder structure
