+++ flask_fullstack/base/__init__.py
@@ -1,0 +1,2 @@
+from .interfaces import Identifiable, UserRole
+from .mixins import AbstractAbortMixin, DatabaseSearcherMixin, JWTAuthorizerMixin
