+++ flask_fullstack/base/interfaces.py
